[NFC] Add a new Intrinsics.cpp file for intrinsic code
Add new file Intrinsics.cpp and move `lookupLLVMIntrinsicByName`
to that file.
